Changed int to unsigned short.

+++ BoardBuilder/BoardBuilder/board.h
@@ -7,16 +7,16 @@
 
 class Board {
 public:
-	bool setHeight(int height);
-	bool setWidth(int width);
+	bool setHeight(unsigned short height);
+	bool setWidth(unsigned short width);
 	bool checkFit(Word word);
 	bool addWord(Word word);
-	std::pair<int, int> getHeightLimits() { return heightLimits; }
-	std::pair<int, int> getWidthLimits() { return widthLimits; }
-	std::pair<int, int> getSize() { return size; }
+	std::pair<unsigned short, unsigned short> getHeightLimits() { return heightLimits; }
+	std::pair<unsigned short, unsigned short> getWidthLimits() { return widthLimits; }
+	std::pair<unsigned short, unsigned short> getSize() { return size; }
 private:
-	std::pair<int, int> heightLimits = std::make_pair(2,20);
-	std::pair<int, int> widthLimits = std::make_pair(2, 20);
-	std::pair<int, int> size;
+	std::pair<unsigned short, unsigned short> heightLimits = std::make_pair(2,20);
+	std::pair<unsigned short, unsigned short> widthLimits = std::make_pair(2, 20);
+	std::pair<unsigned short, unsigned short> size;
 	std::vector<Word> words;
 };
